Fix overview user number tallying error
As it was, the overview user wouldn't be counted if the first JSON
object member didn't have the overview match but the subsequent one did.
Fix this bug by searching all the members for the match.  The user with
potential overview match is not excluded if the first member doesn't
have the match.